ai-model/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
torch
torchvision
transformers
opencv-python
scipy
